requirements.txt
flask==2.0.1
anthropic==0.8.0
google-generativeai==0.3.2
openai==1.3.0
python-dotenv==0.19.0
requests==2.31.0
